+++ crates/ton_core/src/traits.rs
@@ -1,3 +1,3 @@
-pub mod emulator_provider;
+pub mod emulation_provider;
 pub mod state_provider;
 pub mod tlb;